+++ test_install.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
-yarn run build
 /usr/share/weewx/wee_extension --uninstall=weewx-wdc
 /usr/share/weewx/wee_extension --install=.
+/usr/share/weewx/wee_reports